chore(omnibus) go packages renaming
