+++ omnibus/config/software/datadog-agent-prepare.rb
@@ -8,7 +8,7 @@ description "steps required to preprare the build"
 default_version "1.0.0"
 
 license "Apache-2.0"
-license_file "../datadog-dogstatsd/src/github.com/DataDog/datadog-agent/LICENSE"
+license_file "../datadog-dogstatsd/src/github.com/StackVista/stackstate-agent/LICENSE"
 skip_transitive_dependency_licensing true
 
 build do
