+++ omnibus/config/software/datadog-agent.rb
@@ -17,7 +17,7 @@ license "Apache-2.0"
 license_file "../LICENSE"
 
 source path: '..'
-relative_path 'src/github.com/DataDog/datadog-agent'
+relative_path 'src/github.com/StackVista/stackstate-agent'
 
 build do
   # set GOPATH on the omnibus source dir for this software
