+++ omnibus/config/software/datadog-dogstatsd.rb
@@ -11,7 +11,7 @@ license_file "LICENSE"
 skip_transitive_dependency_licensing true
 
 source path: '..'
-relative_path 'src/github.com/DataDog/datadog-agent'
+relative_path 'src/github.com/StackVista/stackstate-agent'
 
 build do
   # set GOPATH on the omnibus source dir for this software
